Chapter 7: World

+++ ray-tracer/Helper.cs
@@ -156,7 +156,7 @@ namespace ray_tracer
         public static Ray Ray(Tuple origin, Tuple direction) => new Ray(origin, direction);
         public static Sphere Sphere() => new Sphere();
 
-        public static Intersection Intersection(int t, Sphere s) => new Intersection(t, s);
+        public static Intersection Intersection(double t, Sphere s) => new Intersection(t, s);
         public static Intersections Intersections() => new Intersections();
         public static Intersections Intersections(params Intersection[] intersections) => new Intersections(intersections);
